% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggers.R
\name{parse_ibutton_list}
\alias{parse_ibutton_list}
\title{Internal function that parses an individual logger data block from a multilogger iButton file}
\usage{
parse_ibutton_list(x)
}
\arguments{
\item{x}{list}
}
\value{
a data.frame
}
\description{
Internal function that parses an individual logger data block from a multilogger iButton file
}
