% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loggers.R
\name{plot.microclim}
\alias{plot.microclim}
\title{Plot microclimate logger output}
\usage{
\method{plot}{microclim}(x, ...)
}
\arguments{
\item{x}{a microclim object}

\item{...}{further arguments to plot}
}
\value{
a plot
}
\description{
Plot microclimate logger output
}

