% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggers.R
\name{read_ibutton_csv}
\alias{read_ibutton_csv}
\title{Read iButton Hygrochron files}
\usage{
read_ibutton_csv(csv_file, parse_name = NULL)
}
\arguments{
\item{csv_file}{input path}

\item{parse_name}{function that tries to extract metadata from the file name}
}
\value{
a data.frame
}
\description{
Read iButton Hygrochron files
}

