% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggers.R
\name{read_ibutton_single_csv}
\alias{read_ibutton_single_csv}
\title{Read iButton Hygrochron single-logger files}
\usage{
read_ibutton_single_csv(csv_file, parse_name = NULL,
  excel_origin = "1899-12-30")
}
\arguments{
\item{csv_file}{input path}

\item{parse_name}{function that tries to extract metadata from the file name}
}
\value{
a data.frame
}
\description{
Function to read csv files containing data dumps of individual iButtons. Accounting for possible corruption in the date column.
}

