% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggers.R
\name{read_inkbird_txt}
\alias{read_inkbird_txt}
\title{Read Ink-Bird THC-4 data logger textfile}
\usage{
read_inkbird_txt(txt_file, parse_name = NULL, tz = NA)
}
\arguments{
\item{txt_file}{input path}

\item{parse_name}{function that tries to extract metadata from the file name}

\item{tz}{string a timezone designation that is compatible with the Olson time zones. See ?timezones for more details.}
}
\value{
a microclim object
}
\description{
Read Ink-Bird THC-4 data logger textfile
}
\section{Warning}{

Temperature data are assumed to be in units degrees Celsius.
}

