% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggers.R
\name{read_usb_logger}
\alias{read_usb_logger}
\title{Read Ecuadorean USB loggers}
\usage{
read_usb_logger(csv_file, parse_name = NULL)
}
\arguments{
\item{parse_name}{}
}
\description{
Read Ecuadorean USB loggers
}

