% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggers.R
\name{read_hobo_csv}
\alias{read_hobo_csv}
\title{Read HOBO loggers}
\usage{
read_hobo_csv(csv_file)
}
\arguments{
\item{csv_file}{path to input csv file}
}
\value{
a data.frame
}
\description{
Read HOBO loggers
}

