% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggers.R
\name{read_hobo_csv}
\alias{read_hobo_csv}
\title{Read HOBO loggers}
\usage{
read_hobo_csv(csv_file, strip_logger_events = TRUE)
}
\arguments{
\item{csv_file}{path to input csv file}

\item{strip_logger_events}{logical, defaults to TRUE. Remove events relating to connecting/disconnecting the coupler and return only the environmental variables}
}
\value{
a data.frame
}
\description{
Read HOBO loggers
}

