% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggers.R
\name{read_ibutton_single_csv}
\alias{read_ibutton_single_csv}
\title{Read iButton Hygrochron single-logger files}
\usage{
read_ibutton_single_csv(csv_file, parse_name = NULL,
  excel_origin = "1899-12-30")
}
\arguments{
\item{csv_file}{input path}

\item{parse_name}{function that tries to extract metadata from the file name}

\item{excel_origin}{origin date for dates encoded as Excel numerical date. Defaults to "1899-12-30", but may have to be changed to "1904-01-01" for certain files. See https://datapub.cdlib.org/2014/04/10/abandon-all-hope-ye-who-enter-dates-in-excel/}
}
\value{
a microclim object
}
\description{
Function to read csv files containing data dumps of individual iButtons. Accounting for possible corruption in the date column.
}

